Make weather prediction
Make a prediction using xgboost based on weather info of the top 10 most frequent cities. This model had an r-squared accuracy of about 0.07 on known test data.
I imagine the accuracy will be lower on the unknown data as it will include cities not in the top 10 list, so weather data for them won't be available. If I see 0.05 on that I think the model will have performed well.